safari-534.53-branch/LayoutTests/media/video-buffered-expected.txt

EXPECTED (video.buffered.length == '0') OK
TEST(video.buffered.start(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.start(0)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(0)) THROWS(DOMException.INDEX_SIZE_ERR) OK
EVENT(canplaythrough)
EXPECTED (video.buffered.length == '1') OK
EXPECTED (video.buffered.start(0) == '0') OK
EXPECTED (video.buffered.end(0) > '0') OK
TEST(video.buffered.start(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.start(1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
END OF TEST
